Use devious tricks to shrink image below 100MB
This took quite a bit of fiddling. The image size is now 90.6MB.

I started with this recipe:

https://blog.realkinetic.com/building-minimal-docker-containers-for-python-applications-37d0272c52f3

This gave me an image size around 102MB. I evaluated image size like so:

    $ docker build . -t json-head
    $ docker images | grep json-head

This is too big! I want less than 100MB. So I ran the image locally and
attached a shell to it like so:

    $ docker run -d -p 8006:8006 json-head
    d99d3434ef7c...
    $ docker exec -it d99d3434ef7c sh

Now that I have a shell into the container I can investigate:

    # du -m /usr/local | sort -n -r | head
    96  /usr/local
    95  /usr/local/lib
    83  /usr/local/lib/python3.6
    36  /usr/local/lib/python3.6/site-packages
    17  /usr/local/lib/python3.6/site-packages/uvloop
    15  /usr/local/lib/python3.6/config-3.6m-x86_64-linux-gnu

This shows me the largest directories measured in MB.

After some digging around, I found some enormous *.c and *.pyd and *.pxd files
that had been left behind by libraries such as uvloop and aiohttp. I altered
the Dockerfile to delete these as part of the build process, and then had it
delete any __pycache__ folders as well.